pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "perpfunding"
version = "0.1.0"
description = "Crypto spot-perp funding arbitrage scanner"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27,<0.28",
    "streamlit>=1.35,<2.0",
    "pandas>=2.2,<3.0",
    "tabulate>=0.9,<1.0",
    "rich>=13.0,<14.0",
    "pyyaml>=6.0,<7.0",
    "python-dotenv>=1.0,<2.0",
    "websockets>=12.0,<14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
]

[project.scripts]
arb = "arb.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["arb"]
